1、修改数据库实体，以重新对应数据库中角色、用户、部门三者之间关系 2、修改所有相关DAO和Server中的对象 3、修正插入和更新数据时，字段数不能动态调整的问题（需要把可能不插入的字段类型设置为对象类型）

+++ src/main/java/siso/edu/cn/entity/DeviceCmdEntity.java
@@ -109,7 +109,7 @@ public class DeviceCmdEntity {
     public int hashCode() {
         int result = (int) (id ^ (id >>> 32));
         result = 31 * result + (cmd != null ? cmd.hashCode() : 0);
-        result = 31 * result + isSend;
+        result = 31 * result + (int) isSend;
         result = 31 * result + (int) (deviceId ^ (deviceId >>> 32));
         result = 31 * result + (createTime != null ? createTime.hashCode() : 0);
         result = 31 * result + (sendTime != null ? sendTime.hashCode() : 0);
